decouple UndefinedVariableResolver
