+++ rules/Php56/Rector/FunctionLike/AddDefaultValueForUndefinedVariableRector.php
@@ -7,27 +7,15 @@ namespace Rector\Php56\Rector\FunctionLike;
 use PhpParser\Node;
 use PhpParser\Node\Expr\Array_;
 use PhpParser\Node\Expr\ArrayDimFetch;
-use PhpParser\Node\Expr\ArrowFunction;
 use PhpParser\Node\Expr\Assign;
-use PhpParser\Node\Expr\AssignRef;
-use PhpParser\Node\Expr\Cast\Unset_ as UnsetCast;
 use PhpParser\Node\Expr\Closure;
-use PhpParser\Node\Expr\List_;
 use PhpParser\Node\Expr\Variable;
-use PhpParser\Node\FunctionLike;
 use PhpParser\Node\Stmt;
 use PhpParser\Node\Stmt\ClassMethod;
 use PhpParser\Node\Stmt\Expression;
-use PhpParser\Node\Stmt\Foreach_;
 use PhpParser\Node\Stmt\Function_;
-use PhpParser\Node\Stmt\Global_;
-use PhpParser\Node\Stmt\Static_;
-use PhpParser\Node\Stmt\StaticVar;
-use PhpParser\Node\Stmt\Unset_;
-use PhpParser\NodeTraverser;
-use PHPStan\Analyser\Scope;
 use Rector\Core\Rector\AbstractRector;
-use Rector\NodeTypeResolver\Node\AttributeKey;
+use Rector\Php56\NodeAnalyzer\UndefinedVariableResolver;
 use Symplify\RuleDocGenerator\ValueObject\CodeSample\CodeSample;
 use Symplify\RuleDocGenerator\ValueObject\RuleDefinition;
 
@@ -40,9 +28,14 @@ use Symplify\RuleDocGenerator\ValueObject\RuleDefinition;
 final class AddDefaultValueForUndefinedVariableRector extends AbstractRector
 {
     /**
-     * @var string[]
+     * @var string
      */
-    private array $definedVariables = [];
+    private const ALREADY_ADDED_VARIABLE_NAMES = 'already_added_variable_names';
+
+    public function __construct(
+        private UndefinedVariableResolver $undefinedVariableResolver
+    ) {
+    }
 
     public function getRuleDefinition(): RuleDefinition
     {
@@ -91,74 +84,31 @@ CODE_SAMPLE
      */
     public function refactor(Node $node): ?Node
     {
-        $this->definedVariables = [];
+        $undefinedVariableNames = $this->undefinedVariableResolver->resolve($node);
 
-        $undefinedVariables = $this->collectUndefinedVariableScope($node);
-        if ($undefinedVariables === []) {
+        // avoids adding same variable multiple tiemes
+        $alreadyAddedVariableNames = (array) $node->getAttribute(self::ALREADY_ADDED_VARIABLE_NAMES);
+        $undefinedVariableNames = array_diff($undefinedVariableNames, $alreadyAddedVariableNames);
+
+        if ($undefinedVariableNames === []) {
             return null;
         }
 
         $variablesInitiation = [];
-        foreach ($undefinedVariables as $undefinedVariable) {
-            if (in_array($undefinedVariable, $this->definedVariables, true)) {
-                continue;
-            }
-
-            $value = $this->isArray($undefinedVariable, (array) $node->stmts)
+        foreach ($undefinedVariableNames as $undefinedVariableName) {
+            $value = $this->isArray($undefinedVariableName, (array) $node->stmts)
                 ? new Array_([])
                 : $this->nodeFactory->createNull();
 
-            $variablesInitiation[] = new Expression(new Assign(new Variable($undefinedVariable), $value));
+            $assign = new Assign(new Variable($undefinedVariableName), $value);
+            $variablesInitiation[] = new Expression($assign);
         }
+
+        $node->setAttribute(self::ALREADY_ADDED_VARIABLE_NAMES, $undefinedVariableNames);
 
         $node->stmts = array_merge($variablesInitiation, (array) $node->stmts);
 
         return $node;
-    }
-
-    /**
-     * @return string[]
-     */
-    private function collectUndefinedVariableScope(ClassMethod | Function_ | Closure $node): array
-    {
-        $undefinedVariables = [];
-
-        $this->traverseNodesWithCallable((array) $node->stmts, function (Node $node) use (&$undefinedVariables): ?int {
-            // entering new scope - break!
-            if ($node instanceof FunctionLike && ! $node instanceof ArrowFunction) {
-                return NodeTraverser::STOP_TRAVERSAL;
-            }
-
-            if ($node instanceof Foreach_) {
-                // handled above
-                $this->collectDefinedVariablesFromForeach($node);
-                return NodeTraverser::DONT_TRAVERSE_CURRENT_AND_CHILDREN;
-            }
-
-            if (! $node instanceof Variable) {
-                return null;
-            }
-
-            if ($this->shouldSkipVariable($node)) {
-                return null;
-            }
-
-            /** @var string $variableName */
-            $variableName = $this->getName($node);
-
-            // defined 100 %
-            /** @var Scope $scope */
-            $scope = $node->getAttribute(AttributeKey::SCOPE);
-            if ($scope->hasVariableType($variableName)->yes()) {
-                return null;
-            }
-
-            $undefinedVariables[] = $variableName;
-
-            return null;
-        });
-
-        return array_unique($undefinedVariables);
     }
 
     /**
@@ -175,90 +125,5 @@ CODE_SAMPLE
 
             return $this->isName($node->var, $undefinedVariable);
         });
-    }
-
-    private function collectDefinedVariablesFromForeach(Foreach_ $foreach): void
-    {
-        $this->traverseNodesWithCallable($foreach->stmts, function (Node $node): void {
-            if ($node instanceof Assign || $node instanceof AssignRef) {
-                if (! $node->var instanceof Variable) {
-                    return;
-                }
-
-                $variableName = $this->getName($node->var);
-                if ($variableName === null) {
-                    return;
-                }
-
-                $this->definedVariables[] = $variableName;
-            }
-        });
-    }
-
-    private function shouldSkipVariable(Variable $variable): bool
-    {
-        $parentNode = $variable->getAttribute(AttributeKey::PARENT_NODE);
-        if (! $parentNode instanceof Node) {
-            return true;
-        }
-
-        if ($parentNode instanceof Global_) {
-            return true;
-        }
-
-        if ($parentNode instanceof Node &&
-            (
-                $parentNode instanceof Assign || $parentNode instanceof AssignRef || $this->isStaticVariable(
-                    $parentNode
-                )
-            )) {
-            return true;
-        }
-
-        if ($parentNode instanceof Unset_ || $parentNode instanceof UnsetCast) {
-            return true;
-        }
-
-        // list() = | [$values] = defines variables as null
-        if ($this->isListAssign($parentNode)) {
-            return true;
-        }
-
-        $nodeScope = $variable->getAttribute(AttributeKey::SCOPE);
-        if (! $nodeScope instanceof Scope) {
-            return true;
-        }
-
-        $variableName = $this->getName($variable);
-
-        // skip $this, as probably in outer scope
-        if ($variableName === 'this') {
-            return true;
-        }
-
-        return $variableName === null;
-    }
-
-    private function isStaticVariable(Node $parentNode): bool
-    {
-        // definition of static variable
-        if ($parentNode instanceof StaticVar) {
-            $parentParentNode = $parentNode->getAttribute(AttributeKey::PARENT_NODE);
-            if ($parentParentNode instanceof Static_) {
-                return true;
-            }
-        }
-
-        return false;
-    }
-
-    private function isListAssign(Node $node): bool
-    {
-        $parentNode = $node->getAttribute(AttributeKey::PARENT_NODE);
-        if ($parentNode instanceof List_) {
-            return true;
-        }
-
-        return $parentNode instanceof Array_;
     }
 }
